server/sample_repository/requirements.txt
# --- Core Orchestration ---
langgraph
langchain-core

# --- LLM Providers (Phase 1) ---
langchain-groq
langchain-google-genai

# --- Environment Management ---
python-dotenv
fastapi==0.111.1
uvicorn[standard]==0.23.2

aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
dataclasses-json==0.6.7
deepeval==3.8.9
distro==1.9.0
execnet==2.1.2
feedparser==6.0.12
filetype==1.2.0
firecrawl-py==4.19.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.67.0
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-groq==1.1.2
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.3
openai==2.28.0
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
pandas==3.0.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyfiglet==1.0.4
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.12.1
PyMuPDF==1.27.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
sentry-sdk==2.54.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
tabulate==0.9.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
